@@ -1,14 +1,20 @@
 
-using Newtonsoft.Json;
 using System.Collections.Generic;
 
-namespace com.pb.shippingapi.model
+namespace PitneyBowes.Developer.ShippingApi.Model
 {
-    public class Customs
+    public class Customs : ICustoms
     {
-        [JsonProperty("customsInfo")]
-        public CustomsInfo CustomsInfo { get;set;}
-        [JsonProperty("customsItems")]
-        public IEnumerable<CustomsItems> CustomsItems {get;set;}
+        public Customs()
+        {
+            CustomsItems = new List<CustomsItems>();
+        }
+
+        virtual public ICustomsInfo CustomsInfo { get; set;}
+        virtual public IEnumerable<ICustomsItems> CustomsItems { get; set; }
+        virtual public ICustomsItems AddCustomsItems(ICustomsItems c)
+        {
+            return ModelHelper.AddToEnumerable<ICustomsItems, CustomsItems>(c, () => CustomsItems, (x) => CustomsItems = x);
+        }
     }
 }
