@@ -2,17 +2,22 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 using System.Collections.Generic;
 
-namespace com.pb.shippingapi.model
+namespace PitneyBowes.Developer.ShippingApi.Model
 {
-    public class SpecialServices
+    public class SpecialServices : ISpecialServices
     {
-        [JsonProperty("specialServiceId")]
-        [JsonConverter(typeof(StringEnumConverter))]
-        public USPSSpecialServiceCodes SpecialServiceId { get; set;}
-        [JsonProperty("inputParameters")]
-        public IEnumerable<Parameter> InputParameters { get;set;}
-       [JsonProperty("fee")]
-        public decimal Fee { get; set;}
+        public SpecialServices()
+        {
+        }
 
+        virtual public USPSSpecialServiceCodes SpecialServiceId { get; set;}
+
+        virtual public IEnumerable<IParameter> InputParameters { get; set; }
+        virtual public void AddParameter(IParameter p)
+        {
+            ModelHelper.AddToEnumerable<IParameter, Parameter>(p, () => InputParameters, (x) => InputParameters = x);
+        }
+        virtual public decimal Fee { get; set;}
     }
+
 }
